requirements.txt
psycopg2-binary

Django>=2.0

django-crispy-forms

django_bootstrap-modal-forms

django-widget-tweaks